idle initialization

+++ Rectify11Installer/Core/Installer.cs
@@ -39,7 +39,7 @@ namespace Rectify11Installer.Core
             // Get all patches
             Patches patches = PatchesParser.GetAll();
             PatchesPatch[] ok = patches.Items;
-            int i = 0;
+            int i = 1;
             bool newpath = false;
             string newhardlink;
             foreach (PatchesPatch patch in ok)
@@ -49,20 +49,22 @@ namespace Rectify11Installer.Core
                     if (items == patch.Mui)
                     {
                         newpath = false;
-                        int number = (i * 100) / ok.Length;
+                        int number = (i /InstallOptions.iconsList.Count) * 100;
                         frm.InstallerProgress = "Patching " + patch.Mui + " (" + number + "%)";
                         if (patch.HardlinkTarget.Contains("%lang%"))
                         {
                             newhardlink = patch.HardlinkTarget.Replace(@"%lang%", Path.Combine(Variables.sys32Folder, CultureInfo.CurrentCulture.Name));
                             if (!Directory.Exists(Path.Combine(Variables.r11Folder, "backup", patch.Mui)))
                                 Directory.CreateDirectory(Path.Combine(Variables.r11Folder, "backup", patch.Mui));
-                            File.Copy(newhardlink, Path.Combine(Variables.r11Folder, "backup", patch.Mui, patch.Mui));
+                            if (!File.Exists(Path.Combine(Variables.r11Folder, "backup", patch.Mui, patch.Mui)))
+                                File.Copy(newhardlink, Path.Combine(Variables.r11Folder, "backup", patch.Mui, patch.Mui));
                         }
                         else if (patch.HardlinkTarget.Contains("mun"))
                         {
                             if (!Directory.Exists(Path.Combine(Variables.r11Folder, "backup", patch.Mui)))
                                 Directory.CreateDirectory(Path.Combine(Variables.r11Folder, "backup", patch.Mui));
-                            File.Copy(patch.HardlinkTarget, Path.Combine(Variables.r11Folder, "backup", patch.Mui, patch.Mui));
+                            if (!File.Exists(Path.Combine(Variables.r11Folder, "backup", patch.Mui, patch.Mui)))
+                                File.Copy(patch.HardlinkTarget, Path.Combine(Variables.r11Folder, "backup", patch.Mui, patch.Mui));
                         }
                         i++;
                     }
